[SPLIT] Splitting tests in order to avoid problems on travis
