+++ .travis.yml
@@ -39,8 +39,14 @@ env:
   - TESTS="1" ODOO_REPO="odoo/odoo" EXCLUDE="cb_medical_guard,cb_test"
   - TESTS="1" ODOO_REPO="OCA/OCB" INCLUDE="cb_medical_guard"
   - TESTS="1" ODOO_REPO="odoo/odoo" INCLUDE="cb_medical_guard"
-  - TESTS="1" ODOO_REPO="OCA/OCB" INCLUDE="cb_test"
-  - TESTS="1" ODOO_REPO="odoo/odoo" INCLUDE="cb_test"
+  - TESTS="1" ODOO_REPO="odoo/odoo" INCLUDE="cb_test_constrains"
+  - TESTS="1" ODOO_REPO="odoo/odoo" INCLUDE="cb_test_document"
+  - TESTS="1" ODOO_REPO="odoo/odoo" INCLUDE="cb_test_invoicing"
+  - TESTS="1" ODOO_REPO="odoo/odoo" INCLUDE="cb_test_laboratory"
+  - TESTS="1" ODOO_REPO="odoo/odoo" INCLUDE="cb_test_payment"
+  - TESTS="1" ODOO_REPO="odoo/odoo" INCLUDE="cb_test_practitioner"
+  - TESTS="1" ODOO_REPO="odoo/odoo" INCLUDE="cb_test_sale"
+  - TESTS="1" ODOO_REPO="odoo/odoo" INCLUDE="cb_test_trigger"
 
 
 install:
