+++ cb_test_invoicing/__manifest__.py
@@ -1,0 +1,18 @@
+# Copyright 2017 Creu Blanca
+# Copyright 2017 Eficent Business and IT Consulting Services, S.L.
+# License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl.html).
+
+{
+    'name': 'CB Testing Sales',
+    'version': '11.0.1.0.0',
+    'author': 'Eficent, Creu Blanca, Odoo Community Association (OCA)',
+    'depends': [
+        'cb_test',
+    ],
+    'data': [
+    ],
+    'website': 'https://github.com/OCA/vertical-medical',
+    'license': 'LGPL-3',
+    'installable': True,
+    'auto_install': False,
+}
